Support the old binary format for Histogram values
Since the change to the format of the Histogram value introduces a
breaking change, it won't be possible to roll this out without downtime.
To work around this, I just added back the logic to parse the value as a
tdigest if it couldn't parse the new format.

I added a test for this behavior, and an accompanying benchmark.  On my
machine, there doesn't seem to be any significant slowdown if local
Veneur's are still sending the old wire format.