Fix behavior in proxy with no destinations
This fixes a bug in the proxies where it will try to forward metrics to
an invalid host if there are no potential destinations. The proxy was
not correctly checking if the hash ring returned a error.

The change proposed here makes the entire ProxyMetrics call fail if the
hash ring returns an error, which I think is reasonable behavior.
It could potentially just ignore that metric, but if that happened I
think it would also make sense to log an error.  Doing that however
would make logs incredibly spammy in the normal case when the proxy is
receving large "import" calls.  By just exiting early out of
ProxyMetrics and failing, it can log just once.