Serialize histogram values as protobufs
Unfortunately, it seems like trying to gob-decode a Histogram value
first into a new `HistoValue` and then falling back to decoding just the
digest is too slow.  After deploying the new version of the global
Veneur's, they started crashing with out of memory errors due to a small
number of Worker goroutines getting choked doing gob-decoding.

Using protobuf's is much faster in this situation, which I confirmed
with the benchmarks written for `(*Histo).Combine` and
`(*Histo).Export`.