Fix merge conflicts with gRPC changes
This change updates the logic for aggregating histogram metrics globally
to start to work with the changes to add a gRPC endpoint. For now, it
just fixes the existing tests.

The next steps are to update the tests added in this change to use gRPC
(rather than the HTTP endpoint) for forwarding.