Generate gRPC forwarding protobuf packages
I completely forgot to include the commands to generate the various gRPC
and protobuf packages for forwarding (#439, #442) in "generate.go".
This change adds in the "protoc-gen-gogofaster" command to the Travis build,
and updates "generate.go" to build the various packages.

In order for the gotoprotobuf extensions to work, I needed to add the
odd looking include statements to the protoc calls.  Since all the three
different packages depend on the base package with the gogoprotobuf
extention, it was necessary it include it in all of them.